Update for form refactoring
