+++ f2/src/Home.js
@@ -29,7 +29,6 @@ import { LocationAnonymousPage } from './LocationAnonymousPage'
 import { Covid19Page } from './Covid19Page'
 import { RedirectRoute } from './components/route'
 import { WhenDidItHappenPage } from './WhenDidItHappenPage'
-import { TestPage } from './TESTPAGE'
 
 export const Home = () => (
   <Router>
@@ -40,9 +39,6 @@ export const Home = () => (
 
       <Route exact path="/">
         <LandingPage />
-      </Route>
-      <Route exact path="/TEST">
-        <TestPage />
       </Route>
       <Route exact path="/startPage">
         <StartPage />
