+++ f2/src/components/formik/checkbox/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { Form, Row } from 'react-bootstrap'
 
 export const CheckBox = ({ field, form, ...props }) => {
-  console.log(`Rendering checkbox ${props.id}`)
   return (
     <Row className="form-row">
       <Form.Check
