+++ f2/src/components/formik/radio/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { Form, Row } from 'react-bootstrap'
 
 export const Radio = ({ field, form, ...props }) => {
-  console.log(`Rendering radio ${props.id}`)
   return (
     <Row className="form-row">
       <Form.Check
