+++ f2/src/components/formik/test/TESTFORM.js
@@ -1,16 +1,16 @@
 import React from 'react'
 import { Trans } from '@lingui/macro'
-import { useStateValue } from '../utils/state'
+import { useStateValue } from '../../../utils/state'
 import { Form, Container, Row } from 'react-bootstrap'
 import { Formik, FieldArray, Field } from 'formik'
-import { CheckBox } from '../components/formik/checkbox'
-import { Radio } from '../components/formik/radio'
-import { TextArea } from '../components/formik/textArea'
-import { Input } from '../components/formik/input'
-import { FileUpload } from '../components/formik/fileUpload'
-import { DatePicker } from '../components/formik/datePicker'
-import { Error, Info, Warning, Success } from '../components/formik/alert'
-import { NextCancelButtons, SkipButton } from '../components/formik/button'
+import { CheckBox } from '../checkbox'
+import { Radio } from '../radio'
+import { TextArea } from '../textArea'
+import { Input } from '../input'
+import { FileUpload } from '../fileUpload'
+import { DatePicker } from '../datePicker'
+import { Error, Info, Warning, Success } from '../alert'
+import { NextCancelButtons, SkipButton } from '../button'
 
 export const TestForm = (props) => {
   const [data] = useStateValue()
